refactor: extract serializeForInlineScript to @isomer/validators
Move inline script serialization into a shared workspace package so
packages/components and tooling/template can both depend on it without
coupling the template to the full components bundle.
