+++ packages/components/src/utils/index.ts
@@ -15,7 +15,7 @@ export { isPhoneNumber, sanitizePhoneNumber } from "./isPhoneNumber"
 export { isUrl } from "./isUrl"
 export { isExternalUrl } from "./isExternalUrl"
 export { safeJsonParse } from "./safeJsonParse"
-export { serializeForInlineScript } from "./serializeForInlineScript"
+export { serializeForInlineScript } from "@isomer/validators"
 export {
   fetchDgsFileDownloadUrl,
   fetchDgsMetadata,
