+++ packages/validators/.oxlintrc.json
@@ -1,0 +1,4 @@
+{
+  "$schema": "../../node_modules/oxlint/configuration_schema.json",
+  "extends": ["../../node_modules/@isomer/oxlint-config/base.json"]
+}
