+++ packages/validators/src/index.ts
@@ -1,0 +1,1 @@
+export { serializeForInlineScript } from "./serializeForInlineScript"
