+++ packages/validators/src/serializeForInlineScript.ts
