+++ packages/validators/tsconfig.json
@@ -1,0 +1,19 @@
+{
+  "compilerOptions": {
+    "types": ["node", "vitest/globals"],
+    "noEmit": true,
+    "moduleResolution": "bundler",
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "strict": true,
+    "noUncheckedIndexedAccess": true,
+    "skipLibCheck": true,
+    "noFallthroughCasesInSwitch": true,
+    "paths": {
+      "~/*": ["./src/*"]
+    },
+    "lib": ["es2021"]
+  },
+  "include": ["src/**/*.ts"],
+  "exclude": ["node_modules"]
+}
