+++ tooling/template/package.json
@@ -15,6 +15,7 @@
     "typecheck": "tsgo --noEmit"
   },
   "dependencies": {
+    "@isomer/validators": "workspace:*",
     "@opengovsg/isomer-components": "workspace:*",
     "isomorphic-dompurify": "catalog:",
     "jsdom": "catalog:",
